adjusting to new API

+++ src/impala/sema/typetable.h
@@ -57,7 +57,7 @@ private:
     Proxy<T> join(T* tn) { garbage_.push_back(tn); return Proxy<T>(tn); }
 
     struct UniHash {
-        size_t operator () (const Unifiable* u) const { return u->hash(); }
+        uint64_t operator () (const Unifiable* u) const { return u->hash(); }
     };
     struct UniEqual {
         bool operator () (const Unifiable* u1, const Unifiable* u2) const { return u1->equal(u2); }
